symfony poprawki css, kontroli dostepu i user index

+++ app/src/Controller/CommentController.php
@@ -16,6 +16,8 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Component\Security\Core\Security;
 
 /**
  * Class CommentController.
@@ -34,15 +36,24 @@ class CommentController extends AbstractController
     private TranslatorInterface $translator;
 
     /**
+     * Security helper.
+     *
+     * @var Security
+     */
+    private Security $security;
+
+    /**
      * Constructor.
      *
      * @param CommentServiceInterface $commentService Comment service
      * @param TranslatorInterface  $translator  Translator
+     * @param Security  $security    Security helper
      */
-    public function __construct(CommentServiceInterface $commentService, TranslatorInterface $translator)
+    public function __construct(CommentServiceInterface $commentService, TranslatorInterface $translator, Security $security)
     {
         $this->commentService = $commentService;
         $this->translator = $translator;
+        $this->security = $security;
     }
 
     /**
@@ -93,34 +104,42 @@ class CommentController extends AbstractController
     #[Route('/{id}/delete/{element}', name: 'comment_delete', requirements: ['id' => '[1-9]\d*'], methods: 'GET|DELETE')]
     public function delete(Request $request, Comment $comment, Element $element): Response
     {
-        $form = $this->createForm(
-            FormType::class,
-            $comment,
-            [
-                'method' => 'DELETE',
-                'action' => $this->generateUrl('comment_delete', ['id' => $comment->getId(), 'element' => $element->getId()]),
-            ]
-        );
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->commentService->delete($comment);
-
-            $this->addFlash(
-                'success',
-                $this->translator->trans('message.deleted_successfully')
+        if ($this->security->isGranted('ROLE_ADMIN'))
+        {
+            $form = $this->createForm(
+                FormType::class,
+                $comment,
+                [
+                    'method' => 'DELETE',
+                    'action' => $this->generateUrl('comment_delete', ['id' => $comment->getId(), 'element' => $element->getId()]),
+                ]
             );
+            $form->handleRequest($request);
 
-            return $this->redirectToRoute('element_show', array ('id' => $element->getId()));
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->commentService->delete($comment);
+
+                $this->addFlash(
+                    'success',
+                    $this->translator->trans('message.deleted_successfully')
+                );
+
+                return $this->redirectToRoute('element_show', array ('id' => $element->getId()));
+            }
+
+            return $this->render(
+                'comment/delete.html.twig',
+                [
+                    'form' => $form->createView(),
+                    'comment' => $comment,
+                ]
+            );
         }
-
-        return $this->render(
-            'comment/delete.html.twig',
-            [
-                'form' => $form->createView(),
-                'comment' => $comment,
-            ]
+        $this->addFlash(
+            'warning',
+            $this->translator->trans('message.access_denied')
         );
+        return $this->redirectToRoute('element_index');
     }
